refactor: spigot-tg-bridge > paper-telegram-bridge
