+++ README.md
@@ -6,13 +6,13 @@ A fork from [spigot-tg-bridge](https://github.com/kraftwerk28/spigot-tg-bridge) 
 
 ## How to use:
 
-1. Download .jar file from [releases page](https://github.com/kraftwerk28/spigot-tg-bridge/releases), and put it in `plugins/` directory on your server **OR** clone this repo and run `gradle` inside repo's directory.
+1. Download .jar file from [releases page](https://github.com/pbl0/paper-telegram-bridge/releases), and put it in `plugins/` directory on your server **OR** clone this repo and run `gradle` inside repo's directory.
 
 2. If you already have telegram bot, skip this step. Otherwise create it through [BotFather](https://t.me/BotFather). You'll go through step-by-step instructions, give a bot __username__ and most importantly, obtain a bot __token__. Save this token for future use. **Note:** in order to make your bot hear raw text messages (not commands), you must disable [privacy mode](https://core.telegram.org/bots#privacy-mode) option which is on by default. Go through bot's settings: **Bot Settings -> Group Privacy** and click **Turn Off**.
 
 3. Next, you need to tell plugin about your new bot. You can either:
-    - Run Spigot server, plugin will log `"No config file found! Saving default one."`. After that, stop server and proceed to 4th step.
-    - Copy [config.yml](https://raw.githubusercontent.com/pbl0/paper-telegram-bridge/master/src/main/resources/config.yml) to `plugins/SpigotTGBridge/` in your server directory.
+    - Run Paper server, plugin will log `"No config file found! Saving default one."`. After that, stop server and proceed to 4th step.
+    - Copy [config.yml](https://raw.githubusercontent.com/pbl0/paper-telegram-bridge/master/src/main/resources/config.yml) to `plugins/PaperTelegramBridge/` in your server directory.
 
 4. A `config.yml` is just a [valid YAML](https://en.wikipedia.org/wiki/YAML) file, alternative for JSON, but more human-readable.
    Now, take bot's __token__ which you got in 2nd step and paste them into `config.yml`, so it looks like this:
