+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'spigot-tg-bridge'
+rootProject.name = 'paper-telegram-bridge'
