+++ src/main/kotlin/eu/pablob/paper_telegram_bridge/Advancement.kt
@@ -1,3 +1,3 @@
-package org.kraftwerk28.spigot_tg_bridge
+package eu.pablob.paper_telegram_bridge
 
 data class Advancement(val displayTitle: String, val key: String)
