+++ src/main/kotlin/eu/pablob/paper_telegram_bridge/Entities.kt
@@ -1,4 +1,4 @@
-package org.kraftwerk28.spigot_tg_bridge
+package eu.pablob.paper_telegram_bridge
 
 import com.google.gson.annotations.SerializedName as Name
