+++ src/main/kotlin/eu/pablob/paper_telegram_bridge/InventoryRenderer.kt
@@ -1,4 +1,4 @@
-package org.kraftwerk28.spigot_tg_bridge
+package eu.pablob.paper_telegram_bridge
 
 import java.awt.Color
 import java.awt.Graphics2D
