+++ src/main/kotlin/eu/pablob/paper_telegram_bridge/ItemRenderer.kt
@@ -1,4 +1,4 @@
-package org.kraftwerk28.spigot_tg_bridge
+package eu.pablob.paper_telegram_bridge
 
 
 import net.kyori.adventure.text.serializer.plain.PlainTextComponentSerializer
