+++ src/main/kotlin/eu/pablob/paper_telegram_bridge/TextureUtils.kt
@@ -1,4 +1,4 @@
-package org.kraftwerk28.spigot_tg_bridge
+package eu.pablob.paper_telegram_bridge
 
 import java.io.InputStream
 import org.bukkit.inventory.meta.PotionMeta
