+++ src/main/resources/plugin.yml
@@ -1,9 +1,9 @@
 name: PaperTelegramBridge
 version: "0.21.0"
 api-version: "1.21.4"
-main: org.kraftwerk28.spigot_tg_bridge.Plugin
+main: eu.pablob.paper_telegram_bridge.Plugin
 description: Telegram <-> Minecraft communication plugin for Paper/Spigot.
 commands:
   tgbridge_reload:
-    description: "Reload Spigot TG bridge plugin"
+    description: "Reload paper-telegram-bridge plugin"
     usage: /<command>
